requirements.txt
flask==3.0.3
google-generativeai==0.8.2
python-dotenv==1.0.1
gunicorn==22.0.0
pytest==8.2.2
